backend/requirements.txt
# Core Framework
Django>=4.2,<5.0

# Django REST Framework (APIs)
djangorestframework>=3.14

# Authentication (JWT)
djangorestframework-simplejwt>=5.3

# CORS (React ↔ Django)
django-cors-headers>=4.3

# API schema/docs
drf-spectacular>=0.27

# Environment Variables
python-dotenv>=1.0

# Database (PostgreSQL support)
psycopg2-binary>=2.9

# AI (optional at runtime; ai_engine falls back if not configured)
openai>=1.6

# Utilities
requests>=2.31
Pillow>=10.1

# Deployment / Server
gunicorn>=21.2
whitenoise>=6.6

# Security
django-filter>=23.5
